% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixWAS.R
\name{combine_site_results}
\alias{combine_site_results}
\title{Combine results from running mixWAS on each site individually}
\usage{
combine_site_results(
  mixWAS_components,
  q,
  phenotypes = NULL,
  phenotype_index = NULL,
  return_p = T
)
}
\arguments{
\item{mixWAS_components:}{list of `mixWAS_single_site` output of length = # of sites}

\item{q:}{# of phenotypes}

\item{phenotypes:}{Optional list of phenotypes, each a matrix of phenotypes (one per column), with names of phenotypes specified as column names.
If `phenotype_index` is `NULL`, will be used to infer phenotypes. One of `phenotypes` and `phenotype_index`must be specified}

\item{phenotype_index:}{list of vectors giving the index (numeric) of which phenotypes are in each site's matrix. If NULL (default), will be
inferred from phenotype matrix colnames.}

\item{return_p:}{logical, if TRUE return P-values, else return components like score/variance. Default = T}
}
\value{
p-value of aggregate test
}
\description{
Combine results from running mixWAS on each site individually
}
