% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixWAS.R
\name{mixWAS}
\alias{mixWAS}
\title{mixWAS Algorithm for All Sites (if all data can be supplied at once)}
\usage{
mixWAS(
  snps,
  phenotypes,
  covariates,
  covariate_map = NULL,
  phenotype_index = NULL,
  types = NULL
)
}
\arguments{
\item{snps:}{list of snps (one for each site), each a vector of SNPs in [0,1,2]}

\item{phenotypes:}{list of phenotypes, each a matrix of phenotypes (one per column), with names of phenotypes specified as column names}

\item{covariates:}{list of covariates, each a matrix or data frame of covariates}

\item{covariate_map:}{default = NULL, list of covariate matrix build instructions for each site.
If specifying for all sites and just need some sites with complex build instruction (not all phenotypes use all covariates)
then supply a list of length = # of sites, entries = NULL for default behavior, and a data frame like object of instructions if needed.
See `mixWAS_single_site` help page for more information.}

\item{phenotype_index:}{list of vectors giving the index (numeric) of which phenotypes are in each site's matrix. If NULL (default), will be
infered from matrix colnames.}

\item{types:}{list of data types ('continuous', 'binary', 'count'). Default = NULL (infer types).
#' Note that 'count' will never be inferred, only 'binary' or 'continuous'.}
}
\value{
p-value of aggregate test
}
\description{
mixWAS Algorithm for All Sites (if all data can be supplied at once)
}
