% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixWAS.R
\name{mixWAS_single_site}
\alias{mixWAS_single_site}
\title{Compute score vector and covariance matrix for a single site}
\usage{
mixWAS_single_site(
  snps,
  phenotypes,
  covariates,
  covariate_map = NULL,
  types = NULL
)
}
\arguments{
\item{snps:}{vector of SNPs in [0,1,2]}

\item{phenotypes:}{matrix of phenotypes (one per column), with names of phenotypes specified as column names}

\item{covariates:}{matrix or data frame of covariates}

\item{covariate_map:}{Default = NULL. If NULL, function assumes by all covariates are to be used for each phenotype.
If this is not desired behavior, user can supply a data frame with two columns
one called `variable` and a second called `phenotype`. In the variable column is the name of covariates, with phenotypes being specified as
'all' (to use the variable for all phenotypes) or the name of a phenotype. Variables can be entered multiple times if they go to multiple phenotypes (but not all).
A phenotype specific data set of covariates with use 'all' covaraite + pheno-type sepcific covariates.}

\item{types:}{optional vector specifying data types ('continuous', 'binary', 'count'). Default = NULL (phenotype data types will be inferred)
Note that 'count' will never be inferred, only 'binary' or 'continuous'.}
}
\value{
a list with components `score` and `V`, the score vector and covariance matrix for single site
}
\description{
Compute score vector and covariance matrix for a single site
}
