% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{run_hypothesis_tests}
\alias{run_hypothesis_tests}
\title{run_hypothesis_tests from mixWAS components}
\usage{
run_hypothesis_tests(score, V_inv, z, q)
}
\arguments{
\item{score}{score vector from mixWAS intermediate output}

\item{V_inv}{Inverse Varariance Matrix from mixWAS intermediate output}

\item{z}{Standardized Z-scores from mixWAS intermediate output}

\item{q}{# of phenotypes}
}
\value{
p-values of aggregate tests
}
\description{
run_hypothesis_tests from mixWAS components
}
