% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixWAS.R
\name{mixWAS}
\alias{mixWAS}
\title{mixWAS Algorithm for All Sites (if all data can be supplied at once)}
\usage{
mixWAS(
  snps,
  phenotypes,
  covariates,
  covariate_map = NULL,
  phenotype_index = NULL,
  types = NULL,
  parallel_sites = F,
  return_p = T
)
}
\arguments{
\item{snps:}{list of snps (one for each site), each a vector of SNPs in [0,1,2]}

\item{phenotypes:}{list of phenotypes, each a matrix of phenotypes (one per column), with names of phenotypes specified as column names}

\item{covariates:}{list of covariates, each a matrix or data frame of covariates}

\item{covariate_map:}{Default = `NULL`. If `NULL`, function assumes by all covariates are to be used for each phenotype.
If this is not desired behavior, user can supply a list of data frames (one for each site). Each data frame should have two columns one called `variable` and a second called `phenotype`.
In the variable column is the name of covariates, with phenotypes being specified as 'all' (to use the variable for all phenotypes)
or the name of a phenotype. Variables can be entered multiple times if they go to multiple phenotypes (but not all). A phenotype specific data set of covariates with use 'all' covariate + phenotype specific covariates.
See `mixWAS_single_site` help page for more information.}

\item{phenotype_index:}{list of vectors giving the index (numeric) of which phenotypes are in each site's matrix. If NULL (default), will be
inferred from matrix colnames.}

\item{types:}{list of data types ('continuous', 'binary', 'count'). Default = NULL (infer types).
#' Note that 'count' will never be inferred, only 'binary' or 'continuous'.}

\item{parallel_sites:}{logical, if score/variance component computations should be parallelized over sites. Default = FALSE}

\item{return_p:}{logical, if TRUE return P-values, else return components like score/variance. Default = T}
}
\value{
p-value of aggregate test
}
\description{
mixWAS Algorithm for All Sites (if all data can be supplied at once)
}
